requirements.txt
unrar==0.4
requests==2.24.0
beautifulsoup4==4.9.1
python_magic==0.4.18
